requirements.txt
fastapi
uvicorn
ics
python-dateutil
py-moneyed